Separate gpg test utils and the vpicc stuff
