+++ tests/rsa4096-import.rs
@@ -2,16 +2,17 @@
 // SPDX-License-Identifier: LGPL-3.0-only
 #![cfg(all(feature = "vpicc", feature = "rsa4096"))]
 
+mod gpg;
 mod virt;
 
 #[cfg(not(feature = "dangerous-test-real-card"))]
 #[test]
 fn gpg_rsa4096_import() {
-    virt::with_vsc(|| virt::gpg_test_import(virt::KeyAlgo::Rsa4096));
+    virt::with_vsc(|| gpg::gpg_test_import(gpg::KeyAlgo::Rsa4096));
 }
 
 #[cfg(feature = "dangerous-test-real-card")]
 #[test]
 fn gpg_rsa4096_import_hardware() {
-    virt::gpg_test_import(virt::KeyAlgo::Rsa4096);
+    gpg::gpg_test_import(gpg::KeyAlgo::Rsa4096);
 }
